// grading-admin-frontend/apps/admin/src/lib/api/v1.d.ts
export interface components {
    schemas: {
// ...
        Course: {
            /** @example 999-9999-9999-99 */
            id?: string;
            /** @example Fall 2020 */
            term: string;
            /** @example true */
            enabled: boolean;
            /** @example EXCL101 */
            name: string;
            /** @example Fall.2020.EXCL.101 */
            code: string;
            /**
             * Format: int64
             * @example 123456
             */
            canvas_id: number;
            members?: components["schemas"]["CourseMember"][];
            assignments?: components["schemas"]["Assignment"][];
            /** @example [
             *       "fall.2020.excl.101.section.a"
             *     ] */
            sections?: string[];
            late_request_config: components["schemas"]["CourseLateRequestConfig"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}